Step 3 - Launch Projectile
Added more code to Slingshot script that allow for the projectile prefab
to instantiate a projectile whenever the mouse button is held down.
Additionally, the projectile is pulled in relation to the mouse position
and launch position.

When mouse button is released, bool aimingMode is set to false and the
projectile is released at a preset velocity in the inspector.