Step 8 - Final?
For the most part, everything gameplay-wise is working as intended. The
hiccups come in when it comes to UI Text. Unfortunately, the Gibson's
site doesn't offer enough information on this particular bit.

To elaborate, yes, Gibson shows changes in component scripting in C#,
but he lacks to mention the updates to the UI Text in this particular
project. While he does offer a link, it redirects to ApplePicker, which
was leaps and bounds less in-depth in terms of scripting.

I was able to get my hands on the second edition of the book, but when
comparing scripts of the two versions. Gibson takes ENTIRELY different
routes in the scripts. Due to this, I cant simply snip out parts of code
that I want to utilize because I need to go back into each and every
individual script to rename variables, make singletons private to
public, etc.

While it's most definitely enticing to see the project to it's very end,
I know too damn well that picking apart each individual line of code in
order to port over the lines from Gibson's 2nd edition would be too
arduous of a process for me to take on. Rather, it'd be better off to
probably start from scratch with the second edition as the sole
reference.

Hopefully, this comment adds some context to my dillemma and any
shortcomings to this particular assignment I'll own up to. The rug got
pulled out from under me on this one. From now on, I'm going to be using
the 2nd edition as my reference.